rename files to adhere to naming convention
