+++ src/scope/request-resolve-scope.ts
