+++ src/scope/singleton-scope.ts
